refactor: 💡 The client example app uses shared wallet provider
