+++ examples/client/src/components/Offers.tsx
@@ -3,7 +3,7 @@ import { stringify } from 'viem';
 import { utils } from '../../../../src/index.js'; // @windingtree/sdk
 import { OfferData } from '../../../../src/shared/types.js';
 import { RequestQuery, OfferOptions } from '../../../shared/index.js';
-import { centerEllipsis } from '../utils.js';
+import { centerEllipsis } from '../../../react-libs/src/utils/index.js';
 
 interface OffersProps {
   offers?: OfferData<RequestQuery, OfferOptions>[];
