+++ examples/client/src/components/Requests.tsx
@@ -1,7 +1,7 @@
 import { ClientRequestRecord, utils } from '../../../../src/index.js'; // @windingtree/sdk
 import { OfferData } from '../../../../src/shared/types.js';
 import { RequestQuery, OfferOptions } from '../../../shared/index.js';
-import { centerEllipsis } from '../utils.js';
+import { centerEllipsis } from '../../../react-libs/src/utils/index.js';
 
 export type RequestsRegistryRecord = Required<
   ClientRequestRecord<RequestQuery, OfferOptions>
